// apps/frontend/src/app/models/payback-request.model.ts
export interface PaybackRequest {
  gender: string;
  marital_status: string;
  education_level: string;
  employment_status: string;
  loan_purpose: string;
  grade_subgrade: string;
  annual_income: number;
  debt_to_income_ratio: number;
  credit_score: number;
  loan_amount: number;
  interest_rate: number;
}